+Command.notes[], *Command.descriptions[]
A limited level of format control over the descriptions/notes is
now possible.

Though each description/note is splitted into words, a leading '\n'
or '\n ' is allowed to isolate a group of lines for a paragraph,
as well as an leading ' ' as indentation.

For each command, the help info includes:

[Command 'name']
descriptions
options info
notes

Except the first registered command, of which the command name is
not displayed, with the notes displayed after all commands.